+++ pytensor/link/numba/cache.py
@@ -1,64 +1,21 @@
-from collections.abc import Callable
+import weakref
+from hashlib import sha256
 from pathlib import Path
-from tempfile import NamedTemporaryFile
-from typing import Any
 
 from numba.core.caching import CacheImpl, _CacheLocator
 
 from pytensor import config
+from pytensor.graph.basic import Apply
 
 
 NUMBA_PYTENSOR_CACHE_ENABLED = True
 NUMBA_CACHE_PATH = config.base_compiledir / "numba"
 NUMBA_CACHE_PATH.mkdir(exist_ok=True)
-CACHED_SRC_FUNCTIONS = {}
-
-
-def compile_and_cache_numba_function_src(
-    src: str,
-    function_name: str,
-    global_env: dict[Any, Any] | None = None,
-    local_env: dict[Any, Any] | None = None,
-    key: str | None = None,
-) -> Callable:
-    if key is not None:
-        filename = NUMBA_CACHE_PATH / key
-        with filename.open("wb") as f:
-            f.write(src.encode())
-    else:
-        with NamedTemporaryFile(delete=False) as f:
-            filename = f.name
-            f.write(src.encode())
-
-    if global_env is None:
-        global_env = {}
-
-    if local_env is None:
-        local_env = {}
-
-    mod_code = compile(src, filename, mode="exec")
-    exec(mod_code, global_env, local_env)
-
-    res = local_env[function_name]
-    res.__source__ = src  # type: ignore
-
-    if key is not None:
-        CACHED_SRC_FUNCTIONS[res] = key
-    return res
-
-
-def cache_numba_function(
-    fn,
-    key: str | None = None,
-) -> Callable:
-    if key is not None:
-        CACHED_SRC_FUNCTIONS[fn] = key
-    return fn
+CACHED_SRC_FUNCTIONS = weakref.WeakKeyDictionary()
 
 
 class NumbaPyTensorCacheLocator(_CacheLocator):
     def __init__(self, py_func, py_file, hash):
-        # print(f"New locator {py_func=}, {py_file=}, {hash=}")
         self._py_func = py_func
         self._py_file = py_file
         self._hash = hash
@@ -101,3 +58,19 @@ class NumbaPyTensorCacheLocator(_CacheLocator):
 
 
 CacheImpl._locator_classes.insert(0, NumbaPyTensorCacheLocator)
+
+
+def cache_node_key(node: Apply, extra_key="") -> str:
+    op = node.op
+    return sha256(
+        str(
+            (
+                # Op signature
+                (type(op), op._props_dict() if hasattr(op, "_props_dict") else ""),
+                # Node signature
+                tuple((type(inp_type := inp.type), inp_type) for inp in node.inputs),
+                # Extra key given by the caller
+                extra_key,
+            ),
+        ).encode()
+    ).hexdigest()
