+++ pytensor/link/numba/dispatch/extra_ops.py
@@ -4,9 +4,11 @@ from typing import cast
 import numba
 import numpy as np
 
+import pytensor.link.numba.compile
 from pytensor.graph import Apply
+from pytensor.link.numba.compile import get_numba_type
 from pytensor.link.numba.dispatch import basic as numba_basic
-from pytensor.link.numba.dispatch.basic import get_numba_type, numba_funcify
+from pytensor.link.numba.dispatch.basic import numba_funcify
 from pytensor.raise_op import CheckAndRaise
 from pytensor.tensor import TensorVariable
 from pytensor.tensor.extra_ops import (
@@ -24,7 +26,7 @@ from pytensor.tensor.extra_ops import (
 
 @numba_funcify.register(Bartlett)
 def numba_funcify_Bartlett(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def bartlett(x):
         return np.bartlett(numba_basic.to_scalar(x))
 
@@ -49,13 +51,13 @@ def numba_funcify_CumOp(op: CumOp, node: Apply, **kwargs):
     if mode == "add":
         if axis is None or ndim == 1:
 
-            @numba_basic.numba_njit
+            @pytensor.link.numba.compile.numba_njit
             def cumop(x):
                 return np.cumsum(x)
 
         else:
 
-            @numba_basic.numba_njit(boundscheck=False)
+            @pytensor.link.numba.compile.numba_njit(boundscheck=False)
             def cumop(x):
                 out_dtype = x.dtype
                 if x.shape[axis] < 2:
@@ -73,13 +75,13 @@ def numba_funcify_CumOp(op: CumOp, node: Apply, **kwargs):
     else:
         if axis is None or ndim == 1:
 
-            @numba_basic.numba_njit
+            @pytensor.link.numba.compile.numba_njit
             def cumop(x):
                 return np.cumprod(x)
 
         else:
 
-            @numba_basic.numba_njit(boundscheck=False)
+            @pytensor.link.numba.compile.numba_njit(boundscheck=False)
             def cumop(x):
                 out_dtype = x.dtype
                 if x.shape[axis] < 2:
@@ -99,7 +101,7 @@ def numba_funcify_CumOp(op: CumOp, node: Apply, **kwargs):
 
 @numba_funcify.register(FillDiagonal)
 def numba_funcify_FillDiagonal(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def filldiagonal(a, val):
         np.fill_diagonal(a, val)
         return a
@@ -109,7 +111,7 @@ def numba_funcify_FillDiagonal(op, **kwargs):
 
 @numba_funcify.register(FillDiagonalOffset)
 def numba_funcify_FillDiagonalOffset(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def filldiagonaloffset(a, val, offset):
         height, width = a.shape
 
@@ -144,25 +146,25 @@ def numba_funcify_RavelMultiIndex(op, node, **kwargs):
 
     if mode == "raise":
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def mode_fn(*args):
             raise ValueError("invalid entry in coordinates array")
 
     elif mode == "wrap":
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def mode_fn(new_arr, i, j, v, d):
             new_arr[i, j] = v % d
 
     elif mode == "clip":
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def mode_fn(new_arr, i, j, v, d):
             new_arr[i, j] = min(max(v, 0), d - 1)
 
     if node.inputs[0].ndim == 0:
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def ravelmultiindex(*inp):
             shape = inp[-1]
             arr = np.stack(inp[:-1])
@@ -178,7 +180,7 @@ def numba_funcify_RavelMultiIndex(op, node, **kwargs):
 
     else:
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def ravelmultiindex(*inp):
             shape = inp[-1]
             arr = np.stack(inp[:-1])
@@ -217,7 +219,7 @@ def numba_funcify_Repeat(op, node, **kwargs):
 
         ret_sig = get_numba_type(node.outputs[0].type)
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def repeatop(x, repeats):
             with numba.objmode(ret=ret_sig):
                 ret = np.repeat(x, repeats, axis)
@@ -228,13 +230,13 @@ def numba_funcify_Repeat(op, node, **kwargs):
 
         if repeats_ndim == 0:
 
-            @numba_basic.numba_njit
+            @pytensor.link.numba.compile.numba_njit
             def repeatop(x, repeats):
                 return np.repeat(x, repeats.item())
 
         else:
 
-            @numba_basic.numba_njit
+            @pytensor.link.numba.compile.numba_njit
             def repeatop(x, repeats):
                 return np.repeat(x, repeats)
 
@@ -259,7 +261,7 @@ def numba_funcify_Unique(op, node, **kwargs):
 
     if not use_python:
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def unique(x):
             return np.unique(x)
 
@@ -277,7 +279,7 @@ def numba_funcify_Unique(op, node, **kwargs):
         else:
             ret_sig = get_numba_type(node.outputs[0].type)
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def unique(x):
             with numba.objmode(ret=ret_sig):
                 ret = np.unique(x, return_index, return_inverse, return_counts, axis)
@@ -297,17 +299,17 @@ def numba_funcify_UnravelIndex(op, node, **kwargs):
 
     if len(node.outputs) == 1:
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def maybe_expand_dim(arr):
             return arr
 
     else:
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def maybe_expand_dim(arr):
             return np.expand_dims(arr, 1)
 
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def unravelindex(arr, shape):
         a = np.ones(len(shape), dtype=np.int64)
         a[1:] = shape[:0:-1]
@@ -340,7 +342,7 @@ def numba_funcify_Searchsorted(op, node, **kwargs):
 
         ret_sig = get_numba_type(node.outputs[0].type)
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def searchsorted(a, v, sorter):
             with numba.objmode(ret=ret_sig):
                 ret = np.searchsorted(a, v, side, sorter)
@@ -348,7 +350,7 @@ def numba_funcify_Searchsorted(op, node, **kwargs):
 
     else:
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def searchsorted(a, v):
             return np.searchsorted(a, v, side)
 
@@ -360,11 +362,11 @@ def numba_funcify_CheckAndRaise(op, node, **kwargs):
     error = op.exc_type
     msg = op.msg
 
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def check_and_raise(x, *conditions):
         for cond in conditions:
             if not cond:
                 raise error(msg)
         return x
 
-    return check_and_raise, 0
+    return check_and_raise
