+++ pytensor/link/numba/dispatch/linalg/solve/utils.py
@@ -1,9 +1,9 @@
 from scipy import linalg
 
-from pytensor.link.numba.dispatch import basic as numba_basic
+import pytensor.link.numba.compile
 
 
-@numba_basic.numba_njit(inline="always")
+@pytensor.link.numba.compile.numba_njit(inline="always")
 def _solve_check_input_shapes(A, B):
     if A.shape[0] != B.shape[0]:
         raise linalg.LinAlgError("Dimensions of A and B do not conform")
