+++ pytensor/link/numba/dispatch/linalg/utils.py
@@ -6,7 +6,7 @@ from numba.core.extending import overload
 from numba.np.linalg import _copy_to_fortran_order, ensure_lapack
 from numpy.linalg import LinAlgError
 
-from pytensor.link.numba.dispatch import basic as numba_basic
+import pytensor.link.numba.compile
 from pytensor.link.numba.dispatch.linalg._LAPACK import (
     _LAPACK,
     _get_underlying_float,
@@ -14,13 +14,13 @@ from pytensor.link.numba.dispatch.linalg._LAPACK import (
 )
 
 
-@numba_basic.numba_njit(inline="always")
+@pytensor.link.numba.compile.numba_njit(inline="always")
 def _copy_to_fortran_order_even_if_1d(x):
     # Numba's _copy_to_fortran_order doesn't do anything for vectors
     return x.copy() if x.ndim == 1 else _copy_to_fortran_order(x)
 
 
-@numba_basic.numba_njit(inline="always")
+@pytensor.link.numba.compile.numba_njit(inline="always")
 def _trans_char_to_int(trans):
     if trans not in [0, 1, 2]:
         raise ValueError('Parameter "trans" should be one of 0, 1, 2')
@@ -53,7 +53,7 @@ def _check_scipy_linalg_matrix(a, func_name):
         raise numba.TypingError(msg, highlighting=False)
 
 
-@numba_basic.numba_njit(inline="always")
+@pytensor.link.numba.compile.numba_njit(inline="always")
 def _solve_check(n, info, lamch=False, rcond=None):
     """
     Check arguments during the different steps of the solution phase
