+++ pytensor/link/numba/dispatch/scalar.py
@@ -2,12 +2,15 @@ import math
 
 import numpy as np
 
+import pytensor.link.numba.compile
 from pytensor.compile.ops import TypeCastingOp
 from pytensor.graph.basic import Variable
-from pytensor.link.numba.cache import compile_and_cache_numba_function_src
+from pytensor.link.numba.compile import (
+    compile_and_cache_numba_function_src,
+    create_numba_signature,
+)
 from pytensor.link.numba.dispatch import basic as numba_basic
 from pytensor.link.numba.dispatch.basic import (
-    create_numba_signature,
     generate_fallback_impl,
     numba_funcify,
 )
@@ -136,12 +139,12 @@ def {scalar_op_fn_name}({', '.join(input_names)}):
 
     # signature = create_numba_signature(node, force_scalar=True)
 
-    return numba_basic.numba_njit(scalar_op_fn)
+    return pytensor.link.numba.compile.numba_njit(scalar_op_fn)
 
 
 @numba_funcify.register(Switch)
 def numba_funcify_Switch(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def switch(condition, x, y):
         if condition:
             return x
@@ -172,7 +175,7 @@ def numba_funcify_Add(op, node, **kwargs):
     signature = create_numba_signature(node, force_scalar=True)
     nary_add_fn = binary_to_nary_func(node.inputs, "add", "+")
 
-    return numba_basic.numba_njit(signature)(nary_add_fn)
+    return pytensor.link.numba.compile.numba_njit(signature)(nary_add_fn)
 
 
 @numba_funcify.register(Mul)
@@ -180,14 +183,14 @@ def numba_funcify_Mul(op, node, **kwargs):
     signature = create_numba_signature(node, force_scalar=True)
     nary_add_fn = binary_to_nary_func(node.inputs, "mul", "*")
 
-    return numba_basic.numba_njit(signature)(nary_add_fn)
+    return pytensor.link.numba.compile.numba_njit(signature)(nary_add_fn)
 
 
 @numba_funcify.register(Cast)
 def numba_funcify_Cast(op, node, **kwargs):
     dtype = np.dtype(op.o_type.dtype)
 
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def cast(x):
         return numba_basic.direct_cast(x, dtype)
 
@@ -197,7 +200,7 @@ def numba_funcify_Cast(op, node, **kwargs):
 @numba_funcify.register(Identity)
 @numba_funcify.register(TypeCastingOp)
 def numba_funcify_type_casting(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def identity(x):
         return x
 
@@ -206,7 +209,7 @@ def numba_funcify_type_casting(op, **kwargs):
 
 @numba_funcify.register(Clip)
 def numba_funcify_Clip(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def clip(x, min_val, max_val):
         x = numba_basic.to_scalar(x)
         min_scalar = numba_basic.to_scalar(min_val)
@@ -228,7 +231,7 @@ def numba_funcify_Composite(op, node, **kwargs):
 
     _ = kwargs.pop("storage_map", None)
 
-    composite_fn = numba_basic.numba_njit(signature)(
+    composite_fn = pytensor.link.numba.compile.numba_njit(signature)(
         numba_funcify(op.fgraph, squeeze_output=True, **kwargs)
     )
     return composite_fn
@@ -236,7 +239,7 @@ def numba_funcify_Composite(op, node, **kwargs):
 
 @numba_funcify.register(Second)
 def numba_funcify_Second(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def second(x, y):
         return y
 
@@ -245,7 +248,7 @@ def numba_funcify_Second(op, node, **kwargs):
 
 @numba_funcify.register(Reciprocal)
 def numba_funcify_Reciprocal(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def reciprocal(x):
         # TODO FIXME: This isn't really the behavior or `numpy.reciprocal` when
         # `x` is an `int`
@@ -256,7 +259,7 @@ def numba_funcify_Reciprocal(op, node, **kwargs):
 
 @numba_funcify.register(Sigmoid)
 def numba_funcify_Sigmoid(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def sigmoid(x):
         return 1 / (1 + np.exp(-x))
 
@@ -265,7 +268,7 @@ def numba_funcify_Sigmoid(op, node, **kwargs):
 
 @numba_funcify.register(GammaLn)
 def numba_funcify_GammaLn(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def gammaln(x):
         return math.lgamma(x)
 
@@ -274,7 +277,7 @@ def numba_funcify_GammaLn(op, node, **kwargs):
 
 @numba_funcify.register(Log1mexp)
 def numba_funcify_Log1mexp(op, node, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def logp1mexp(x):
         if x < np.log(0.5):
             return np.log1p(-np.exp(x))
@@ -286,7 +289,7 @@ def numba_funcify_Log1mexp(op, node, **kwargs):
 
 @numba_funcify.register(Erf)
 def numba_funcify_Erf(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def erf(x):
         return math.erf(x)
 
@@ -295,7 +298,7 @@ def numba_funcify_Erf(op, **kwargs):
 
 @numba_funcify.register(Erfc)
 def numba_funcify_Erfc(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def erfc(x):
         return math.erfc(x)
 
@@ -306,7 +309,7 @@ def numba_funcify_Erfc(op, **kwargs):
 def numba_funcify_Softplus(op, node, **kwargs):
     out_dtype = np.dtype(node.outputs[0].type.dtype)
 
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def softplus(x):
         if x < -37.0:
             value = np.exp(x)
