+++ pytensor/link/numba/dispatch/scan.py
@@ -4,13 +4,13 @@ import numpy as np
 from numba import types
 from numba.extending import overload
 
+import pytensor.link.numba.compile
 from pytensor import In
 from pytensor.compile.function.types import add_supervisor_to_fgraph
 from pytensor.compile.mode import NUMBA, get_mode
+from pytensor.link.numba.compile import create_arg_string, create_tuple_string
 from pytensor.link.numba.dispatch import basic as numba_basic
 from pytensor.link.numba.dispatch.basic import (
-    create_arg_string,
-    create_tuple_string,
     numba_funcify,
 )
 from pytensor.link.utils import compile_function_src
@@ -97,7 +97,7 @@ def numba_funcify_Scan(op: Scan, node, **kwargs):
     )
     rewriter(fgraph)
 
-    scan_inner_func = numba_basic.numba_njit(numba_funcify(op.fgraph))
+    scan_inner_func = pytensor.link.numba.compile.numba_njit(numba_funcify(op.fgraph))
 
     outer_in_names_to_vars = {
         (f"outer_in_{i}" if i > 0 else "n_steps"): v for i, v in enumerate(node.inputs)
@@ -442,4 +442,4 @@ def scan({", ".join(outer_in_names)}):
 
     scan_op_fn = compile_function_src(scan_op_src, "scan", {**globals(), **global_env})
 
-    return numba_basic.numba_njit(scan_op_fn, boundscheck=False)
+    return pytensor.link.numba.compile.numba_njit(scan_op_fn, boundscheck=False)
