+++ pytensor/link/numba/dispatch/tensor_basic.py
@@ -2,9 +2,13 @@ from textwrap import indent
 
 import numpy as np
 
-from pytensor.link.numba.cache import compile_and_cache_numba_function_src
+import pytensor.link.numba.compile
+from pytensor.link.numba.compile import (
+    compile_and_cache_numba_function_src,
+    create_tuple_string,
+)
 from pytensor.link.numba.dispatch import basic as numba_basic
-from pytensor.link.numba.dispatch.basic import create_tuple_string, numba_funcify
+from pytensor.link.numba.dispatch.basic import numba_funcify
 from pytensor.link.utils import unique_name_generator
 from pytensor.tensor.basic import (
     Alloc,
@@ -58,7 +62,7 @@ def allocempty({", ".join(shape_var_names)}):
     )
 
     return (
-        numba_basic.numba_njit(alloc_fn),
+        pytensor.link.numba.compile.numba_njit(alloc_fn),
         hash_from_code(alloc_def_src),
     )
 
@@ -107,7 +111,7 @@ def alloc(val, {", ".join(shape_var_names)}):
     )
 
     return (
-        numba_basic.numba_njit(alloc_fn),
+        pytensor.link.numba.compile.numba_njit(alloc_fn),
         hash_from_code(alloc_def_src),
     )
 
@@ -116,7 +120,7 @@ def alloc(val, {", ".join(shape_var_names)}):
 def numba_funcify_ARange(op, **kwargs):
     dtype = np.dtype(op.dtype)
 
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def arange(start, stop, step):
         return np.arange(
             start.item(),
@@ -130,20 +134,20 @@ def numba_funcify_ARange(op, **kwargs):
 
 @numba_funcify.register(Join)
 def numba_funcify_Join(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def join(axis, *tensors):
         return np.concatenate(tensors, axis.item())
 
-    return join, 0
+    return join
 
 
 @numba_funcify.register(Split)
 def numba_funcify_Split(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def split(tensor, axis, indices):
         return np.split(tensor, np.cumsum(indices)[:-1], axis=axis.item())
 
-    return split, 0
+    return split
 
 
 @numba_funcify.register(ExtractDiag)
@@ -153,7 +157,7 @@ def numba_funcify_ExtractDiag(op, node, **kwargs):
 
     if node.inputs[0].type.ndim == 2:
 
-        @numba_basic.numba_njit(inline="always")
+        @pytensor.link.numba.compile.numba_njit(inline="always")
         def extract_diag(x):
             out = np.diag(x, k=offset)
 
@@ -168,7 +172,7 @@ def numba_funcify_ExtractDiag(op, node, **kwargs):
         leading_dims = (slice(None),) * axis1
         middle_dims = (slice(None),) * (axis2 - axis1 - 1)
 
-        @numba_basic.numba_njit
+        @pytensor.link.numba.compile.numba_njit
         def extract_diag(x):
             if offset >= 0:
                 diag_len = min(x.shape[axis1], max(0, x.shape[axis2] - offset))
@@ -193,7 +197,7 @@ def numba_funcify_ExtractDiag(op, node, **kwargs):
 def numba_funcify_Eye(op, **kwargs):
     dtype = np.dtype(op.dtype)
 
-    @numba_basic.numba_njit(inline="always")
+    @pytensor.link.numba.compile.numba_njit(inline="always")
     def eye(N, M, k):
         return np.eye(
             numba_basic.to_scalar(N),
@@ -231,21 +235,23 @@ def makevector({", ".join(input_names)}):
         "makevector",
         {**globals(), **global_env},
     )
-    return numba_basic.numba_njit(makevector_fn), hash_from_code(makevector_def_src)
+    return pytensor.link.numba.compile.numba_njit(makevector_fn), hash_from_code(
+        makevector_def_src
+    )
 
 
 @numba_funcify.register(TensorFromScalar)
 def numba_funcify_TensorFromScalar(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def tensor_from_scalar(x):
         return np.array(x)
 
-    return tensor_from_scalar, 0
+    return tensor_from_scalar
 
 
 @numba_funcify.register(ScalarFromTensor)
 def numba_funcify_ScalarFromTensor(op, **kwargs):
-    @numba_basic.numba_njit
+    @pytensor.link.numba.compile.numba_njit
     def scalar_from_tensor(x):
         return x.item()
