+++ pytensor/link/numba/linker.py
@@ -17,7 +17,7 @@ class NumbaLinker(JITLinker):
         if self.vm:
             return fn
         else:
-            from pytensor.link.numba.dispatch.basic import numba_njit
+            from pytensor.link.numba.compile import numba_njit
 
             jitted_fn = numba_njit(fn, no_cpython_wrapper=False, no_cfunc_wrapper=False)
             return jitted_fn
