+++ tests/link/numba/test_tensor_basic.py
@@ -280,7 +280,7 @@ def test_ExtractDiag(val, offset):
 )
 @pytest.mark.parametrize("reverse_axis", (False, True))
 def test_ExtractDiag_exhaustive(k, axis1, axis2, reverse_axis):
-    from pytensor.link.numba.dispatch.basic import numba_njit
+    from pytensor.link.numba.compile import numba_njit
 
     if reverse_axis:
         axis1, axis2 = axis2, axis1
